+++ metadata.rb
@@ -4,6 +4,6 @@ maintainer_email 'daniel.paul.searles@gmail.com'
 license 'MIT'
 description 'Installs/Configures squaresurf_influxdb'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version '1.0.0'
+version '1.1.0'
 
 supports 'ubuntu'