+++ resources/database.rb
@@ -28,3 +28,4 @@ actions(:create, :delete)
 default_action(:create)
 
 attribute(:database, kind_of: String, name_attribute: true)
+attribute(:options, kind_of: Hash, default: {})